[fix] entrypoint.prd.sh

+++ entrypoint.prd.sh
@@ -6,21 +6,22 @@ rm -f /myapp/tmp/pids/server.pid
 
 cd /myapp
 
-echo "PORT " $PORT " is used"
+if [[ -e ./src ]]; then
+  # on first build
+  echo "run npm install"
+  npm install
 
-echo "run npm install"
-npm install
+  echo "run npm upgrade --force"
+  # npm update --force
 
-echo "run npm upgrade --force"
-# npm update --force
+  echo "npm run build"
+  npm run build
 
-echo "npm run build"
-npm run build
+  echo "remove node files"
+  rm -rf ./src
 
-echo "remove node files"
-rm -rf ./src
-
-rails db:migrate
+  rails db:migrate
+fi
 
 # コンテナーのプロセスを実行する。（Dockerfile 内の CMD に設定されているもの。）
 exec "$@"
